requirements.txt
opencv-python==4.6.0.66
pyserial==3.5
paddleocr==2.5.0.3